requirements.txt
torch
torchvision
torchaudio
transformers
peft
pillow
qwen-vl-utils
flash-attn

psycopg[binary]
qdrant-client

pymupdf
pytesseract
gdown
dotenv